python-agent/requirements.txt
livekit-agents>=1.2.15
livekit-plugins-deepgram>=1.2.15
livekit-plugins-elevenlabs>=1.2.15
livekit-plugins-silero>=1.2.15
livekit-plugins-google>=1.2.15
python-dotenv>=1.0.0
aiohttp>=3.9.0